fix build on GCC 7

+++ mailbox.c
@@ -37,8 +37,9 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <unistd.h>
 #include <sys/mman.h>
 #include <sys/ioctl.h>
-#include <sys/types.h>
 #include <sys/stat.h>
+#include <sys/sysmacros.h>
+#include <sys/types.h>
 
 #include "mailbox.h"
 
@@ -68,10 +69,10 @@ void *mapmem(uint32_t base, uint32_t size) {
 
 void *unmapmem(void *addr, uint32_t size) {
     uint32_t pagemask = ~0UL ^ (getpagesize() - 1);
-    uint32_t baseaddr = (uint32_t)addr & pagemask;
+    uint32_t baseaddr = (uint32_t)(intptr_t)addr & pagemask;
     int s;
-    
-    s = munmap((void *)baseaddr, size);
+
+    s = munmap((void *)(intptr_t)baseaddr, size);
     if (s != 0) {
         perror("munmap error\n");
     }
@@ -190,7 +191,7 @@ uint32_t mem_unlock(int file_desc, uint32_t handle) {
    return p[5];
 }
 
-uint32_t execute_code(int file_desc, uint32_t code, uint32_t r0, uint32_t r1, 
+uint32_t execute_code(int file_desc, uint32_t code, uint32_t r0, uint32_t r1,
                       uint32_t r2, uint32_t r3, uint32_t r4, uint32_t r5) {
     int i=0;
     uint32_t p[32];
